Introduce tests for dash_ag_grid function
